+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Added
 
+- Add `AuthSanitizer::Loader.load` for isolated loading without defining top-level `Auth`
+
 ### Changed
 
 ### Deprecated